refactor(model-server): implement health endpoint via base class
Moves the health endpoint over to the operative API (as it is mainly
used for operations) and implements it using a generated base class
instead of a manual declaration that can drift from the spec.
